Require user to specify what version is changing.
This is for both `update` and `archive`.

Also:

+ Remove `setUnlined*` functions.

These were used to change only the instance of a card without changing
any of the other data on the card. The intent was to minimize what gets
overridden if two applications write new versions of the same card at
about the same time.

Trying to handle this by minimizing the damage seems like the wrong
answer though -- with the changes to `update` and `archive` now conflicts
must be handled explicitly.

+ remove `set*` functions.

These are covered fine by `new` and `update`.

+ Remove the `CardMag` type alias.

+ Change README slightly.